Rename jumpBackInCell and put it it's in own file
